tests: Extract FirebaseEmulatorContainer to new file
